% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.r
\name{calculate_weights}
\alias{calculate_weights}
\title{Prediction-domain adaptive weighting based on raking}
\usage{
calculate_weights(
  tpoints,
  modeldomain = NULL,
  predpoints = NULL,
  pointwise_error = NULL,
  samplesize = 1000,
  sampling = "regular",
  balance_by = 0.2,
  shrink_lambda = 0.2,
  ...
)
}
\arguments{
\item{tpoints}{data.frame or sf object containing the predictor values at the training points.}

\item{modeldomain}{SpatRaster containing the predictors. Not needed when predpoints are supplied.}

\item{predpoints}{data.frame or sf object containing the predictor values at the prediction points.
Only needed if no modeldomain is supplied.}

\item{pointwise_error}{vector of the error estimates for every training points (aligned with the training points). Optional.
If supplied, the weights will be applied to the pointwise error, and an weighted error will be returned.}

\item{samplesize}{numeric. How many points in the modeldomain should be sampled as prediction points?
Only required if modeldomain is used instead of predpoints.}

\item{sampling}{character. How to draw prediction points from the modeldomain? See \code{sf::st_sample}.
Only required if modeldomain is used instead of predpoints.}

\item{balance_by}{Numeric step size for quantile binning.}

\item{shrink_lambda}{Shrinkage parameter applied after calibration raking.}
}
\value{
A list with elements:
\describe{
\item{weights}{Named list of weight object for TWCV.}
\item{weighted_error}{vector of the weighted error. Only if pointwise_error is supplied.}
\item{unsupported_flag}{1 if some quintiles are not supported by the training data. 0 otherwise.}
\item{unsupported_vars}{Vector containing the names of variables with quintiles not supported by the training data.}
\item{training_bal}{data.frame containing the discretized predictors for the training data.}
\item{prediction_margins}{data.frame containing the frequency of the discretized predictors for the prediction data}
}
}
\description{
Prediction-domain adaptive weighting based on raking
}
