% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\alias{plot.twcv}
\title{Plot the calibration results}
\usage{
\method{plot}{twcv}(x, pointwise_error = NULL, ...)
}
\arguments{
\item{x}{An object of type \emph{twcv}.}

\item{pointwise_error}{Optional. Either a \code{pe} object (from
\code{\link[=tw_pointwise_error]{tw_pointwise_error()}}) or a numeric vector of pointwise errors. If a
numeric vector is supplied, it is assumed to be aligned with the weights by
position. If a \code{pe} object is supplied, errors are aligned to the weights by
ID. If \code{NULL}, only the calibration plot is returned.}

\item{...}{other arguments.}
}
\description{
Generic plot function for twcv class
}
\author{
Jan Linnenbrink
}
