% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionMatching-package.R
\docType{package}
\name{PredictionMatching-package}
\alias{PredictionMatching}
\alias{PredictionMatching-package}
\title{PredictionMatching: Prediction-domain Adaptive Weighting for Cross-Validation Evaluation}
\description{
Implements prediction-domain adaptive weighting via raking to reduce evaluation bias from preferential training samples. The method reweights cross-validation errors so they better reflect the target prediction domain's properties, supporting more accurate spatial machine learning model evaluation (Brenning and Suesse 2026 <10.48550/ARXIV.2603.29981>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JanLinnenbrink/PredictionMatching}
  \item Report bugs at \url{https://github.com/JanLinnenbrink/PredictionMatching/issues}
}

}
\author{
\strong{Maintainer}: Jan Linnenbrink \email{jan.linnenbrink@uni-muenster.de}

Authors:
\itemize{
  \item Jan Linnenbrink \email{jan.linnenbrink@uni-muenster.de}
}

}
\keyword{internal}
