% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_error.r
\name{weighted_error_stats}
\alias{weighted_error_stats}
\title{Weighted error statistics}
\usage{
weighted_error_stats(weights, pointwise_error)
}
\arguments{
\item{weights}{A \code{twcv} object, or a numeric vector of weights.}

\item{pointwise_error}{A vector of the error for each training point based on
cross-validation (obs - pred).}
}
\value{
A named numeric vector of weighted error statistics
(bias, mse, rmse, mae).
}
\description{
Weighted error statistics
}
