% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\alias{plot.twcv}
\title{Plot the calibration results}
\usage{
\method{plot}{twcv}(x, pointwise_error = NULL, ...)
}
\arguments{
\item{x}{An object of type \emph{twcv}.}

\item{pointwise_error}{vector of errors for each training point.}

\item{...}{other arguments.}
}
\description{
Generic plot function for twcv class
}
\author{
Jan Linnenbrink
}
