% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_calculate_weights.r
\name{tw_calculate_weights}
\alias{tw_calculate_weights}
\title{Prediction-domain adaptive weighting based on raking}
\usage{
tw_calculate_weights(
  tpoints,
  modeldomain = NULL,
  predpoints = NULL,
  samplesize = 1000,
  sampling = "regular",
  balance_by = 0.2,
  shrink_lambda = 0.2
)
}
\arguments{
\item{tpoints}{data.frame or sf object containing the predictor values at the training points.}

\item{modeldomain}{SpatRaster containing the predictors. Not needed when predpoints are supplied.}

\item{predpoints}{data.frame or sf object containing the predictor values at the prediction points.
Only needed if no modeldomain is supplied.}

\item{samplesize}{numeric. How many points in the modeldomain should be sampled as prediction points?
Only required if modeldomain is used instead of predpoints.}

\item{sampling}{character. How to draw prediction points from the modeldomain? See \code{sf::st_sample}.
Only required if modeldomain is used instead of predpoints.}

\item{balance_by}{Numeric step size for quantile binning.}

\item{shrink_lambda}{Shrinkage parameter applied after calibration raking.}
}
\value{
A list with elements:
\describe{
\item{weights}{Named list of weight object for TWCV.}
\item{ids}{Vector of ids for the training points, useful for aligning weights with pointwise errors in weighted_error_stats.}
\item{unsupported_flag}{1 if some quintiles are not supported by the training data. 0 otherwise.}
\item{unsupported_vars}{Vector containing the names of variables with quintiles not supported by the training data.}
\item{ess}{effective sample size.}
\item{ess_ratio}{Ratio of the effective sample size. Ranges from 0 (weights are highly concentrated) to 1.}
\item{training_bal}{data.frame containing the discretized predictors for the training data.}
\item{prediction_margins}{data.frame containing the frequency of the discretized predictors for the prediction data}
}
}
\description{
Prediction-domain adaptive weighting based on raking
}
\examples{
\dontrun{
library(sf)
library(terra)
library(ggplot2)
library(CAST)

data(splotdata)
splotdata <- splotdata[splotdata$Country == "Chile",]

predictors <- c("bio_1", "bio_4", "bio_5", "bio_6",
               "bio_8", "bio_9", "bio_12", "bio_13",
               "bio_14", "bio_15", "elev")

trainDat <- sf::st_drop_geometry(splotdata)
predictors_sp <- terra::rast(system.file("extdata", "predictors_chile.tif",package="CAST"))
terra::plot(predictors_sp[["bio_1"]])
terra::plot(vect(splotdata), add = T)

w <- tw_calculate_weights(tpoints = trainDat[,predictors], modeldomain = predictors_sp)
plot(w)
}

}
